Added better error messages
When an error happens at the client side it will show a small modal for 10seconds with the status and a nice error message.
When nog logged in, or session wrong, the user is automatically logged out correct.
Also fixed a wrong response code (was 2020, now 200) in the logout function of admin panel
Also fixed a wrong check in the middelware.js of the admin panel