Worked mostly on game controls in app
Added the components (without any style or html)
Added a way to hide nav-bar on game page
Added a way so you cannot enter the game-page without a game running or joined

Worked a little bit on uWebsockets
Added a "joinMidGame" parameter so games have that option.

Going to work further on game controls in app now.